requirements.txt
click==8.1.3
gitdb==4.0.10
GitPython==3.1.30
prompt-toolkit==1.0.14
Pygments==2.14.0
PyInquirer==1.0.3
regex==2022.10.31
semantic-version==2.10.0
six==1.16.0
smmap==5.0.0
wcwidth==0.2.5
